requirements.txt
requests
openai
python-dotenv
langchain
langchain-community
langchain-google-community
google-api-python-client
